refactor: 입력로직을 제외하고 테스트하기 쉽게 만들기 위해 race 컨트롤러 생성
